requirements.txt
Flask
PyYAML
openai
python-dotenv
pandas
